+++ resources/views/bootstrap-4/form-radio.blade.php
@@ -13,9 +13,13 @@
         @if($checked)
             checked="checked"
         @endif
+
+        @if($label && !$attributes->get('id'))
+            id="{{ $id() }}"
+        @endif
     />
 
-   <x-form-label :label="$label" :for="$name" class="form-check-label" />
+   <x-form-label :label="$label" :for="$attributes->get('id') ?: $id()" class="form-check-label" />
 
     {!! $help ?? null !!}