+++ resources/views/bootstrap-4/form-textarea.blade.php
@@ -1,11 +1,15 @@
 <div class="form-group">
-    <x-form-label :label="$label" :for="$name" />
+    <x-form-label :label="$label" :for="$attributes->get('id') ?: $id()" />
 
     <textarea
         @if($isWired())
             wire:model="{{ $name }}"
         @else
             name="{{ $name }}"
+        @endif
+
+        @if($label && !$attributes->get('id'))
+            id="{{ $id() }}"
         @endif
 
         {!! $attributes->merge(['class' => 'form-control ' . ($hasError($name) ? 'is-invalid' : '')]) !!}