+++ tests/Feature/views/bootstrap-label-for.blade.php
@@ -1,0 +1,4 @@
+<x-form>
+    <x-form-input name="input" label="Input" />
+    <x-form-textarea name="textarea" label="Textarea" id="text_b" />
+</x-form>